Cambio Id de int a long, Añadi ServiceImpl
Cambie los id de las clases en users y en contenido el tipo de dato,
espero no haber salteado algun constructor (igual le va hacer cast, ya
que va pedir int y guardar un int en un long, creo).
Comence a añadir la implementacion de los servicios pero aun hay mucho
camino por recorrer.

+++ src/main/java/py/edu/uc/lp3/content/Pelicula.java
@@ -6,7 +6,7 @@ public class Pelicula extends Video{
 	int emision; //fecha en la que salio
 	
 	
-	public Pelicula(String nombre, String genero, int id, List<SitioReview> ratings, int emision) {
+	public Pelicula(String nombre, String genero, long id, List<SitioReview> ratings, int emision) {
 		super(nombre, genero, id, ratings);
 		this.emision = emision;
 	}
